@@ -1,56 +1,45 @@
 {
-    "name": "travel-app",
+    "name": "travel-weather",
     "version": "1.0.0",
-    "description": "",
+    "description": "Final project for Udacity Front End Web Developer Nanodegree",
+    "main": "index.js",
     "scripts": {
-        "test": "jest",
-        "start": "node src/server/index.js",
-        "dev": "webpack-dev-server --config webpack.dev.js --open",
-        "build": "webpack --config webpack.dev.js",
-        "build-prod": "webpack --config webpack.prod.js"
+      "test": "jest",
+      "start": "node src/server/start.js",
+      "dev": "webpack-dev-server --config webpack.dev.js --open",
+      "build": "webpack --config webpack.prod.js"
     },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/harshitagupta30/travel-app.git"
-    },
-    "author": "harshitagupta310@gmail.com",
+    "author": "JeffCad",
     "license": "ISC",
-    "bugs": {
-        "url": "https://github.com/harshitagupta30/travel-app/issues"
-    },
-    "homepage": "https://github.com/harshitagupta30/travel-app#readme",
     "dependencies": {
-        "body-parser": "^1.19.0",
-        "cors": "^2.8.5",
-        "express": "^4.17.1",
-        "webpack": "^4.43.0",
-        "webpack-cli": "^3.3.12"
+      "body-parser": "^1.19.0",
+      "cors": "^2.8.5",
+      "countdown": "^2.6.0",
+      "dotenv": "^8.2.0",
+      "express": "^4.17.1",
+      "node-fetch": "^2.6.1"
     },
     "devDependencies": {
-        "@babel/core": "^7.10.3",
-        "@babel/plugin-transform-runtime": "^7.10.3",
-        "@babel/preset-env": "^7.10.3",
-        "@babel/runtime": "^7.10.3",
-        "axios": "^0.19.2",
-        "babel-loader": "^8.1.0",
-        "babel-polyfill": "^6.26.0",
-        "clean-webpack-plugin": "^3.0.0",
-        "css-loader": "^3.6.0",
-        "dotenv": "^8.2.0",
-        "file-loader": "^6.0.0",
-        "html-loader": "^1.1.0",
-        "html-webpack-plugin": "^4.3.0",
-        "jest": "^26.1.0",
-        "jquery": "^3.5.1",
-        "mini-css-extract-plugin": "^0.9.0",
-        "node-fetch": "^2.6.0",
-        "node-mocks-http": "^1.8.1",
-        "node-sass": "^4.14.1",
-        "optimize-css-assets-webpack-plugin": "^5.0.3",
-        "sass-loader": "^8.0.2",
-        "style-loader": "^1.2.1",
-        "supertest": "^4.0.2",
-        "webpack-dev-server": "^3.11.0",
-        "workbox-webpack-plugin": "^5.1.3"
+      "@babel/core": "^7.11.6",
+      "@babel/preset-env": "^7.11.5",
+      "babel-loader": "^8.1.0",
+      "clean-webpack-plugin": "^3.0.0",
+      "css-loader": "^4.3.0",
+      "favicons-webpack-plugin": "^4.2.0",
+      "file-loader": "^6.1.0",
+      "html-loader": "^1.3.1",
+      "html-webpack-plugin": "^4.5.0",
+      "jest": "^26.4.2",
+      "mini-css-extract-plugin": "^0.11.3",
+      "node-sass": "^4.14.1",
+      "optimize-css-assets-webpack-plugin": "^5.0.4",
+      "sass-loader": "^10.0.2",
+      "style-loader": "^1.3.0",
+      "supertest": "^5.0.0",
+      "terser-webpack-plugin": "^4.2.2",
+      "webpack": "^4.44.2",
+      "webpack-cli": "^3.3.12",
+      "webpack-dev-server": "^3.11.0",
+      "workbox-webpack-plugin": "^5.1.4"
     }
-}
+  }
