@@ -1,21 +1,22 @@
-const path = require('path');
-const webpack = require('webpack');
-const HtmlWebPackPlugin = require('html-webpack-plugin');
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const HtmlWebPackPlugin = require('html-webpack-plugin')
+const { CleanWebpackPlugin } = require('clean-webpack-plugin')
+const FaviconsWebpackPlugin = require('favicons-webpack-plugin')
 
 module.exports = {
     entry: './src/client/index.js',
     mode: 'development',
     devtool: 'source-map',
-    target: 'node',
     stats: 'verbose',
+    // Outputs all functions to a client library
+    // Necessary only if more than 1 JavaScript file
     output: {
         libraryTarget: 'var',
         library: 'Client'
     },
     module: {
-        rules: [{
-                test: '/\.js$/',
+        rules: [
+            {
+                test: /\.js$/,
                 exclude: /node_modules/,
                 loader: 'babel-loader'
             },
@@ -24,28 +25,33 @@ module.exports = {
                 use: ['style-loader', 'css-loader', 'sass-loader']
             },
             {
-                test: /\.(jpg|png|svg|jpg|gif)$/,
-                loader: 'file-loader'
+                test: /\.png$|\.jpg$/,
+                loader: 'file-loader',
+                options: {
+                    name: '[name].[ext]',
+                    outputPath: 'icons'
+                }
+            },
+            {
+                test: /\.html$/,
+                loader: 'html-loader'
             },
         ]
     },
     plugins: [
         new HtmlWebPackPlugin({
-            template: './src/client/views/index.html',
-            filename: './index.html',
-        }),
-        new HtmlWebPackPlugin({
-            template: './src/client/views/saved-trips.html',
-            filename: './saved-trips.html',
+            template: "./src/client/views/index.html",
+            filename: "./index.html",
         }),
         new CleanWebpackPlugin({
             // Simulate the removal of files
-            dry: true,
+            dry: false,
             // Write Logs to Console
             verbose: true,
             // Automatically remove all unused webpack assets on rebuild
             cleanStaleWebpackAssets: true,
             protectWebpackAssets: false
-        })
+        }),
+        new FaviconsWebpackPlugin('./src/client/icons/favicon.ico'),
     ]
 }
