@@ -1,33 +1,42 @@
-const path = require('path');
-const webpack = require('webpack');
-const HtmlWebPackPlugin = require('html-webpack-plugin');
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const HtmlWebPackPlugin = require('html-webpack-plugin')
+const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
-const WorkboxPlugin = require('workbox-webpack-plugin');
+const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
+const TerserPlugin = require('terser-webpack-plugin')
+const WorkboxPlugin = require('workbox-webpack-plugin')
+const FaviconsWebpackPlugin = require('favicons-webpack-plugin')
 
 module.exports = {
     entry: './src/client/index.js',
+    mode: 'production',
+    // Outputs all functions to a client library
+    // Necessary only if more than 1 JavaScript file
     output: {
         libraryTarget: 'var',
-        library: 'Client',
-        path: path.resolve(__dirname, 'dist'),
-        filename: 'bundle.min.js'
+        library: 'Client'
     },
-    mode: 'production',
     module: {
-        rules: [{
-                test: '/\.js$/',
+        rules: [
+            {
+                test: /\.js$/,
                 exclude: /node_modules/,
-                loader: "babel-loader"
+                loader: 'babel-loader'
             },
             {
                 test: /\.scss$/,
                 use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader']
             },
             {
-                test: /\.(jpg|png|svg|jpg|gif)$/,
-                loader: 'file-loader'
+                test: /\.png$|\.jpg$/,
+                loader: 'file-loader',
+                options: {
+                    name: '[name].[ext]',
+                    outputPath: 'icons'
+                }
+            },
+            {
+                test: /\.html$/,
+                loader: 'html-loader'
             },
         ]
     },
@@ -36,23 +45,23 @@ module.exports = {
             template: "./src/client/views/index.html",
             filename: "./index.html",
         }),
-        new HtmlWebPackPlugin({
-            template: './src/client/views/saved-trips.html',
-            filename: './saved-trips.html',
-        }),
         new CleanWebpackPlugin({
             // Simulate the removal of files
-            dry: true,
+            dry: false,
             // Write Logs to Console
             verbose: true,
             // Automatically remove all unused webpack assets on rebuild
             cleanStaleWebpackAssets: true,
             protectWebpackAssets: false
         }),
-        new MiniCssExtractPlugin(),
-        new WorkboxPlugin.GenerateSW({
-            clientsClaim: true,
-            skipWaiting: true
-        })
-    ]
+        new MiniCssExtractPlugin({ filename: "[name].css" }),
+        new WorkboxPlugin.GenerateSW(),
+        new FaviconsWebpackPlugin({
+            logo: './src/client/icons/favicon.ico',
+            mode: 'light',
+        }),
+    ],
+    optimization: {
+        minimizer: [new TerserPlugin({}), new OptimizeCSSAssetsPlugin({})],
+    },
 }
